chore: move posthog test instance helper
