+++ functional_tests/feature-flags.test.ts
@@ -1,5 +1,5 @@
 import { v4 } from 'uuid'
-import { createPosthogInstance } from './posthog-instance'
+import { createPosthogInstance } from '../src/__tests__/helpers/posthog-instance'
 import { waitFor } from '@testing-library/dom'
 import { getRequests, resetRequests } from './mock-server'
